% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_case_slurm.R
\name{assign_real_case_tasks}
\alias{assign_real_case_tasks}
\title{Assign tasks to real-case runs with optional shuffling.}
\usage{
assign_real_case_tasks(runs, task_max = 200, shuffle_seed = NULL)
}
\description{
Assign tasks to real-case runs with optional shuffling.
}
\keyword{internal}
