% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_results.R
\name{compute_auprc_single}
\alias{compute_auprc_single}
\title{Compute AUPRC for a single precision-recall curve}
\usage{
compute_auprc_single(precision, recall)
}
\arguments{
\item{precision}{Numeric vector of precision values}

\item{recall}{Numeric vector of recall values}
}
\value{
Single AUPRC value
}
\description{
Compute AUPRC for a single precision-recall curve
}
\keyword{internal}
