% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metrics.R, R/x_matrix_metrics.R
\name{compute_auprc_single_internal}
\alias{compute_auprc_single_internal}
\title{Internal AUPRC computation (single curve)}
\usage{
compute_auprc_single_internal(precision, recall)

compute_auprc_single_internal(precision, recall)
}
\description{
Internal AUPRC computation (single curve)

Internal AUPRC computation (single curve)
}
\keyword{internal}
