% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metrics.R, R/x_matrix_metrics.R
\name{compute_x_metrics_from_file}
\alias{compute_x_metrics_from_file}
\title{Compute X-matrix metrics from a genotype matrix file}
\usage{
compute_x_metrics_from_file(x_path, k_ref = 3, n_samples = 200)

compute_x_metrics_from_file(x_path, k_ref = 3, n_samples = 200)
}
\arguments{
\item{x_path}{Path to file containing X matrix (samples x SNPs).
Supports CSV, TSV, and gzipped versions (.gz)}

\item{k_ref}{Reference number of causal effects}

\item{n_samples}{Number of MC samples for sparse eigenvalue}
}
\value{
Tibble with all metrics

Tibble with all metrics
}
\description{
Compute X-matrix metrics from a genotype matrix file

Compute X-matrix metrics from a genotype matrix file
}
