% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metrics.R
\name{eigs_summaries}
\alias{eigs_summaries}
\title{Compute eigenvalue-based summaries}
\usage{
eigs_summaries(R)
}
\arguments{
\item{R}{Correlation matrix}
}
\value{
Tibble with spectral metrics
}
\description{
Compute eigenvalue-based summaries
}
\keyword{internal}
