% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{execute_dataset_bundle}
\alias{execute_dataset_bundle}
\title{Execute all model runs for a dataset bundle.}
\usage{
execute_dataset_bundle(
  bundle_runs,
  job_config,
  quiet = FALSE,
  buffer_ctx = NULL
)
}
\description{
Execute all model runs for a dataset bundle.
}
\keyword{internal}
