% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metrics.R, R/x_matrix_metrics.R
\name{fit_gam_safe}
\alias{fit_gam_safe}
\title{Safely fit a GAM and extract diagnostics}
\usage{
fit_gam_safe(df, y_var, m1, m2 = NULL, k = 5, interaction = FALSE)

fit_gam_safe(df, y_var, m1, m2 = NULL, k = 5, interaction = FALSE)
}
\description{
Safely fit a GAM and extract diagnostics

Safely fit a GAM and extract diagnostics
}
\keyword{internal}
