% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{generate_data_for_bundle}
\alias{generate_data_for_bundle}
\title{Generate data bundle for a dataset bundle row (shared across model runs).}
\usage{
generate_data_for_bundle(bundle_row, job_config)
}
\description{
Generate data bundle for a dataset bundle row (shared across model runs).
}
\keyword{internal}
