% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{generate_data_for_run}
\alias{generate_data_for_run}
\title{Generate data bundle for a given run_row based on scenario.}
\usage{
generate_data_for_run(run_row, job_config)
}
\description{
Generate data bundle for a given run_row based on scenario.
}
\keyword{internal}
