% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metrics.R, R/x_matrix_metrics.R
\name{get_axis_b_metrics}
\alias{get_axis_b_metrics}
\title{Get Axis B metric names (Identifiability / Factorization Mismatch)}
\usage{
get_axis_b_metrics()

get_axis_b_metrics()
}
\value{
Character vector of Axis B metric names

Character vector of Axis B metric names
}
\description{
Returns metrics that measure how interchangeable SNP neighborhoods are.
High Axis B = many SNPs look alike, SuSiE's factorization may be wrong shape.

Returns metrics that measure how interchangeable SNP neighborhoods are.
High Axis B = many SNPs look alike, SuSiE's factorization may be wrong shape.
}
