% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_results.R
\name{get_use_case_labels}
\alias{get_use_case_labels}
\title{Get use case labels from catalog}
\usage{
get_use_case_labels(use_case_ids = NULL)
}
\arguments{
\item{use_case_ids}{Character vector of use_case_id values to look up}
}
\value{
Named character vector mapping use_case_id to label
}
\description{
Get use case labels from catalog
}
