% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metrics.R, R/x_matrix_metrics.R
\name{get_x_metric_names}
\alias{get_x_metric_names}
\title{Get default list of X-metric names}
\usage{
get_x_metric_names()

get_x_metric_names()
}
\value{
Character vector of metric column names

Character vector of metric column names
}
\description{
Get default list of X-metric names

Get default list of X-metric names
}
