% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{lookup_use_case}
\alias{lookup_use_case}
\title{Retrieve metadata row for a given use case ID.}
\usage{
lookup_use_case(job_config, use_case_id)
}
\description{
Retrieve metadata row for a given use case ID.
}
\keyword{internal}
