% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metrics.R
\name{mid_energy_dist_weighted}
\alias{mid_energy_dist_weighted}
\title{Compute distance-weighted mid-energy
Detects non-local "plaid" LD patterns}
\usage{
mid_energy_dist_weighted(R)
}
\arguments{
\item{R}{Correlation matrix}
}
\value{
Scalar M1_dist value
}
\description{
Compute distance-weighted mid-energy
Detects non-local "plaid" LD patterns
}
\keyword{internal}
