% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metrics.R
\name{mid_energy_M1}
\alias{mid_energy_M1}
\title{Compute mid-energy M1 (linear)
Higher values indicate more mass in ambiguous mid-LD range}
\usage{
mid_energy_M1(R)
}
\arguments{
\item{R}{Correlation matrix}
}
\value{
Scalar M1 value
}
\description{
Compute mid-energy M1 (linear)
Higher values indicate more mass in ambiguous mid-LD range
}
\keyword{internal}
