% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metrics.R
\name{mid_energy_M2}
\alias{mid_energy_M2}
\title{Compute mid-energy M2 (quadratic)
Sharper emphasis on moderate-high mid-LD}
\usage{
mid_energy_M2(R)
}
\arguments{
\item{R}{Correlation matrix}
}
\value{
Scalar M2 value
}
\description{
Compute mid-energy M2 (quadratic)
Sharper emphasis on moderate-high mid-LD
}
\keyword{internal}
