% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metrics.R
\name{mutual_coherence}
\alias{mutual_coherence}
\title{Compute mutual coherence (max off-diagonal |r|)}
\usage{
mutual_coherence(R)
}
\arguments{
\item{R}{Correlation matrix}
}
\value{
Scalar mu value
}
\description{
Compute mutual coherence (max off-diagonal |r|)
}
\keyword{internal}
