% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metrics.R, R/x_matrix_metrics.R
\name{offdiag_means}
\alias{offdiag_means}
\title{Compute off-diagonal means}
\usage{
offdiag_means(R)

offdiag_means(R)
}
\arguments{
\item{R}{Correlation matrix}
}
\value{
Tibble with mean |r| and mean r^2

Tibble with mean |r| and mean r^2
}
\description{
Compute off-diagonal means

Compute off-diagonal means
}
\keyword{internal}
