% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_controls.R
\name{relativize_path}
\alias{relativize_path}
\title{Convert an absolute path to one relative to the repository root.}
\usage{
relativize_path(path, root)
}
\description{
Convert an absolute path to one relative to the repository root.
}
\keyword{internal}
