% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_case_slurm.R
\name{render_real_case_slurm_script}
\alias{render_real_case_slurm_script}
\title{Render SLURM submission script for real-case runs.}
\usage{
render_real_case_slurm_script(job_config, run_task_script)
}
\description{
Render SLURM submission script for real-case runs.
}
\keyword{internal}
