% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_controls.R
\name{render_slurm_script}
\alias{render_slurm_script}
\title{Render a SLURM submission script string.}
\usage{
render_slurm_script(job_config, run_task_script)
}
\arguments{
\item{job_config}{Job configuration list.}

\item{run_task_script}{Path to \code{run_task.R}.}
}
\value{
Character vector of script lines.
}
\description{
Render a SLURM submission script string.
}
\keyword{internal}
