% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_flag}
\alias{resolve_flag}
\title{Convert a logical flag to scalar TRUE/FALSE, applying default when NA.}
\usage{
resolve_flag(x, default = FALSE)
}
\arguments{
\item{x}{Logical input.}

\item{default}{Logical default used when x is NA.}
}
\description{
Convert a logical flag to scalar TRUE/FALSE, applying default when NA.
}
\keyword{internal}
