% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_cases.R
\name{resolve_use_cases}
\alias{resolve_use_cases}
\title{Resolve model specs by id.}
\usage{
resolve_use_cases(ids = NULL)
}
\arguments{
\item{ids}{Character vector.}
}
\value{
Tibble subset of prior specs.
}
\description{
Accepts either prior_spec_id values or use_case_id aliases.
}
