% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_controls.R
\name{summarise_job_config}
\alias{summarise_job_config}
\title{Summarise expected run counts per use case and scenario.}
\usage{
summarise_job_config(job_config)
}
\arguments{
\item{job_config}{Job configuration list.}
}
\value{
Tibble summarising runs.
}
\description{
Summarise expected run counts per use case and scenario.
}
