% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timestamp_utc}
\alias{timestamp_utc}
\title{Format a timestamp in ISO8601 UTC.}
\usage{
timestamp_utc()
}
\value{
Character scalar timestamp.
}
\description{
Format a timestamp in ISO8601 UTC.
}
\keyword{internal}
