% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_cases.R
\name{valid_exploration_for_prior}
\alias{valid_exploration_for_prior}
\title{Validate exploration compatibility with prior specs.}
\usage{
valid_exploration_for_prior(prior_spec_ids = NULL, exploration_ids = NULL)
}
\arguments{
\item{prior_spec_ids}{Character vector of prior_spec_id values.}

\item{exploration_ids}{Character vector of exploration_id values.}
}
\value{
Tibble with columns prior_spec_id, exploration_id, valid.
}
\description{
Validate exploration compatibility with prior specs.
}
