% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_controls.R
\name{write_job_artifacts}
\alias{write_job_artifacts}
\title{Write job configuration and scripts to disk.}
\usage{
write_job_artifacts(job_config, run_task_script)
}
\arguments{
\item{job_config}{Output of \code{\link[=make_job_config]{make_job_config()}}.}

\item{run_task_script}{Path to the general task-runner script.}
}
\value{
List with paths to the serialized artifacts.
}
\description{
Write job configuration and scripts to disk.
}
