% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_case_slurm.R
\name{write_real_case_job_artifacts}
\alias{write_real_case_job_artifacts}
\title{Write real-case job artifacts to disk.}
\usage{
write_real_case_job_artifacts(job_config, run_task_script)
}
\arguments{
\item{job_config}{Output of \code{\link[=build_real_case_job_config]{build_real_case_job_config()}}.}

\item{run_task_script}{Path to task runner script.}
}
\value{
List of artifact paths.
}
\description{
Write real-case job artifacts to disk.
}
